+++ src/Utils/CmsUtils.php
@@ -12,6 +12,15 @@ use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsAnyFilter;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
 
+/**
+ * This class provides utility methods to work with the shopware cms. It has build in caching to prevent
+ * multiple database queries for the same data within one command execution / request.
+ *
+ * This class is designed to be used through the FixtureHelper, using:
+ * ```php
+ * $this->helper->Cms()->……();
+ * ```
+ */
 readonly class CmsUtils
 {
     /**
@@ -24,15 +33,19 @@ readonly class CmsUtils
 
     public function getDefaultCategoryLayout(): ?CmsPageEntity
     {
-        $criteria = (new Criteria())
-            ->addFilter(new EqualsFilter('locked', '1'))
-            ->addFilter(new EqualsAnyFilter('translations.name', ['Default category layout', 'Default listing layout']))
-            ->setLimit(1);
+        return once(function (): ?CmsPageEntity {
+            $criteria = (new Criteria())
+                ->addFilter(new EqualsFilter('locked', '1'))
+                ->addFilter(new EqualsAnyFilter('translations.name', ['Default category layout', 'Default listing layout']))
+                ->setLimit(1);
 
-        $cmsPage = $this->cmsPageRepository
-            ->search($criteria, Context::createDefaultContext())
-            ->first();
+            $criteria->setTitle(sprintf('%s::%s()', __CLASS__, __FUNCTION__));
 
-        return $cmsPage instanceof CmsPageEntity ? $cmsPage : null;
+            $cmsPage = $this->cmsPageRepository
+                ->search($criteria, Context::createDefaultContext())
+                ->first();
+
+            return $cmsPage instanceof CmsPageEntity ? $cmsPage : null;
+        });
     }
 }